Add files via upload
# Todo List

**Task Management System**

## Overview

This repository hosts a simple web-based todo list . It provides basic functionality for adding, deleting, and marking tasks as completed.

## Getting Started

### Prerequisites

- Web browser

### Installation

1. Clone the repository:

   ```bash
   git clone https://github.com/sassante/your-todo-list.git

Open index.html in a web browser.
Usage
Use the "Add Task" button to add a new task to the list.
Tasks are limited to 9; attempting to add more will display an alert.
Use the checkbox to mark a task as completed, which will strike through the task text.
Click the delete icon (🗑) to remove a task.
Contributing
Contributions are welcome! If you have suggestions, improvements, or bug fixes, please feel free to open an issue or submit a pull request.

License
This project is licensed under the MIT License.


Contact
For inquiries, reach out to sandra.bh.asante@gmail.com.